modules/location_ingest/requirements.txt
kafka-python==2.0.2
grpcio-tools==1.51.1
Faker==13.7.0